refactor: [EL-4391] improve the code according to comments
