+++ src/[fsd]/entities/credential-warning/helpers/credentialWarning.helpers.js
