+++ src/[fsd]/entities/credential-warning/helpers/index.js
@@ -1,0 +1,3 @@
+import * as CredentialWarningHelpers from './credentialWarning.helpers.js';
+
+export { CredentialWarningHelpers };
