+++ src/[fsd]/entities/credential-warning/hooks/index.js
@@ -1,0 +1,1 @@
+export * from './useCredentialWarning.hooks';
