+++ src/[fsd]/entities/credential-warning/ui/index.js
@@ -1,0 +1,1 @@
+export { default as CredentialWarningModal } from './CredentialWarningModal';
