--- src/[fsd]/shared/ui/modal/credential-warning/index.js
@@ -1,1 +1,0 @@
-export { CredentialWarningModal } from './CredentialWarningModal';
